minor bugfix dispose online lesson

+++ models/model-classroom.js
@@ -81,10 +81,10 @@ function disposeOnlinelesson (recentLesson) {
     if (fs.existsSync(path.join(__dirname, '../data/classes', recentLesson.group, 'onlinelesson.png'))) {
       fs.unlinkSync(path.join(__dirname, '../data/classes', recentLesson.group, 'onlinelesson.png'));
     }
-    if (recentLesson.docs[0].includes('onlinelesson') && fs.existsSync(path.join(__dirname, '../', recentLesson.chalkboardBg))) {
+    if (recentLesson.docs[0] && recentLesson.docs[0].includes('onlinelesson') && fs.existsSync(path.join(__dirname, '../', recentLesson.chalkboardBg))) {
       fs.unlinkSync(path.join(__dirname, '../', recentLesson.chalkboardBg));
     }
-    if (recentLesson.docs[0].includes('onlinelesson') && fs.existsSync(path.join(__dirname, '../', recentLesson.docs[0]))) {
+    if (recentLesson.docs[0] && recentLesson.docs[0].includes('onlinelesson') && fs.existsSync(path.join(__dirname, '../', recentLesson.docs[0]))) {
       fs.unlinkSync(path.join(__dirname, '../', recentLesson.docs[0]));
     }
   } catch (e) {
